pyproject.toml
[project]
authors = [{ name = "Hiro", email = "hiroto125takeuchi@gmail.com" }]
dependencies = [
    # --- Universal dependencies ---
    "docopt",
    "websocket-client",
    "opencv-python",
    "opencv-contrib-python",
    "numpy",
    "pyyaml",
    "pillow",
    "customtkinter>=5.2.2",
    "sv-ttk>=2.6.1",
    "darkdetect>=0.8.0",
    "loguru>=0.7.3",
    "textual>=6.7.0",
    # "PySide6>=6.10.1" # includes pyside6-essentials and pyside6-addons
    "pyside6-essentials>=6.10.1",
    # "qtwidgets" # for Toggle Switch Widget. Might be able to just take the code directly since it is MIT licensed
    "av>=16.1.0",
    "pydantic>=2.12.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pyobjc-framework-Cocoa; sys_platform == 'darwin'",
]
name = "commander"
requires-python = "==3.12.*"
version = "0.1.0"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.scripts]
commander = "talos:main"
commander-terminal = "talos:terminal_interface"
commander-tk = "talos:tk_interface"

[dependency-groups]
dev = ["pyinstaller>=6.16.0", "textual-dev>=1.8.0"]

[project.optional-dependencies]
mediapipe = [
    "mediapipe>=0.10.32; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "torch>=2.6.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    # --- Platform-Specific Dependencies ---
    # For x86_64 macOS: use specific versions due to compatibility issues
    "mediapipe==0.10.21; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "jax[cpu]==0.4.38; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "torch==2.2.2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
]
yolo = ["ultralytics"]

[tool.uv]
package = true

[tool.setuptools]
py-modules = ["talos"]
include-package-data = true

[tool.setuptools.package-data]
assets = ["**/*"]
config = ["**/*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "assets*", "config*"]
